requirements.txt
streamlit>=1.28.0
deep-translator>=1.11.4
groq>=0.11.0
speechrecognition>=3.10.0
gtts>=2.4.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
# pyaudio: optional, for microphone input. Fails on Windows with Python 3.14 (no pre-built wheel).
# Install separately if needed: pip install -r requirements-voice.txt (requires Python 3.8-3.13 on Windows)
